requirements.txt
Django==3.1.1
djangorestframework==3.12.2
django-cors-headers